# server/openapi.yaml
openapi: 3.0.3
info:
  title: EmailDiet Multi-User SaaS API
  description: |
    Self-hosted and multi-user Gmail optimizer API.
    All operations are non-destructive by default (moving to Gmail TRASH only) and operate on email metadata headers (`From`, `Subject`, `Date`, `List-Unsubscribe`, `List-Unsubscribe-Post`).
    
    ### Authentication & User Isolation
    Protected endpoints require a valid session JWT cookie (`oauth_state` or `auth_token`).
    Every request and database operation is strictly scoped to the authenticated user ID (`req.userId` extracted from the JWT cookie) adhering to a 1:1 user/account model.
  version: 1.0.0
  contact:
    name: EmailDiet Support
    email: support@emaildiet.com
servers:
  - url: http://localhost:3001
    description: Local Development Server
  - url: https://api.emaildiet.com
    description: Production Environment
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP-only JWT session cookie obtained via `/api/auth/callback` or `/api/auth/demo-login`.
...
paths:
...
  /api/labels:
    get:
      tags: [Labels]
      summary: List all app-created and system Gmail labels
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of labels
  /api/labels/{id}:
    delete:
      tags: [Labels]
      summary: Delete an app-created Gmail label
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Label deleted
...
  /api/logs/recent:
    get:
      tags: [Logs]
      summary: Get recent activity audit logs
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Recent activity logs
  /api/logs/metadata:
    get:
      tags: [Logs]
      summary: Get scan timing and profiling history
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Scan timing metadata
  /api/logs/client:
    post:
      tags: [Logs]
      summary: Forward client-side console errors to server terminal
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: string
                  default: info
                message:
                  type: string
                context:
                  type: object
      responses:
        '200':
          description: Client log recorded